% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{close_run_state}
\alias{close_run_state}
\title{Close a run-state (flush and close the log connection)}
\usage{
close_run_state(state)
}
\description{
Close a run-state (flush and close the log connection)
}
\keyword{internal}
